fixing code smells
