+++ caliber/src/main/java/com/revature/caliber/controllers/BootController.java
@@ -65,7 +65,7 @@ public class BootController extends Helper {
 	 * @throws URISyntaxException
 	 *             the uri syntax exception
 	 */
-	//@RequestMapping(value = "/caliber")
+	@RequestMapping(value = "/caliber")
 	public String devHomePage(HttpServletRequest servletRequest, HttpServletResponse servletResponse)
 			throws IOException, URISyntaxException {
 		HttpClient httpClient = HttpClientBuilder.create().build();
@@ -111,7 +111,7 @@ public class BootController extends Helper {
 	 * @throws URISyntaxException
 	 *             the uri syntax exception
 	 */
-	@RequestMapping(value = "/caliber")
+	//@RequestMapping(value = "/caliber")
 	public String getHomePage(HttpServletRequest servletRequest, HttpServletResponse servletResponse)
 			throws IOException, URISyntaxException {
 		HttpClient httpClient = HttpClientBuilder.create().build();
