+++ caliber/src/main/java/com/revature/caliber/security/impl/AuthorizationImpl.java
@@ -68,7 +68,7 @@ public class AuthorizationImpl extends Helper implements Authorization {
 	 * 
 	 * TODO remove @RequestMapping at go-live
 	 */
-	//@RequestMapping("/")
+	@RequestMapping("/")
 	public ModelAndView dummyAuth() {
 		return new ModelAndView(REDIRECT + redirectUrl);
 	}
@@ -80,7 +80,7 @@ public class AuthorizationImpl extends Helper implements Authorization {
 	 * 
 	 * TODO enable at go-live
 	 */
-	@RequestMapping("/")
+	//@RequestMapping("/")
 	public ModelAndView openAuthURI() {
 		return new ModelAndView(
 				REDIRECT +loginURL+ authURL + "?response_type=code&client_id=" + clientId + "&redirect_uri=" + redirectUri);
